[Hunter] cleanup bugged Focused Trickery triggers
For most of the spells the buggy behaviour is the fact that it uses the
*base* spell cost. There are some outliers that either don't contribute
to it at all (Barrage), use the "really base" cost (AMoC which uses the
dbc base cost before the passive aura reduction) or require complete
special handling (AiS).

We can make it a bit more automatic and explicit at the same time by
controlling it through a (maybe) bool on hunter_action_t and generically
triggering it from its execute. Makes the special cases stand out and
removes the boilerplate from most "normal" spells.