[Hunter] AiS Vigil munching on impact
AiS impacts consume Secrets of the Unblinking Vigil buff on impact if
the latter was up on cast.

So in a sequence like MS (TrS+Vigil proc) -> AiS -> MS (TrS+Vigil proc)
the in-flight AiS will munch Vigil from the last MS.

While at it factor out an action state helper to reduce boilerplate
associated with it.