Fix build errors
fmt with C++20 enabled now has compile time checks (if possible) of the format string against parameter types. This can detect formatting problems.

Not sure if this would actually have resulted in problem when formatting the enums, but it can't hurt to explicitly cast them to a integer type as that is the desired way to format them in those places that I adjusted.