[Gear] resonant reservoir initial implementation
The trinket fires a missile than randomly lands around the target.
Once it lands it triggers an expanding ring.
When the ring hits a target, it applies a dot.

Each use of the trinket will cause the next use to fire an additional missile with a delay.
If the dot is refreshed, the duration is refreshed to full 6s but the current tick is NOT clipped.
This means that the cumulative duration of the dot will depend on the randomness of the final application timing. As ticks are not clipped, this will impact the damage of the last partial tick.

TODO: extremely annoying to do as none of these things show up on the combat log
  * confirm delay between repeated launches
  * confirm delay between missile hitting the ground, expanding the ring, and the ring hitting the target