[DBC] switch to floats for random prop data
The fields are doubles because in
  https://github.com/simulationcraft/simc/pull/5799
we had to support both ptr and live spell data and the latter had them
as integers with some values resulting in narrowing conversion errors.

Now that we don't have to care about integers and all of the source
values in client data a 32b floats we can safely use floats like every
other item-related data.