[SV] apply jayeasy's trinket magic
Mileage may vary with Jotun, currently uses jotun AFTER any trinkets, but this may or may not be "ideal". in some scenarios (eg a 252 IQD) using jotun first results in ~+40 dps.
given how jotun clashes with trinket CDs the effort required to properly optimize this for a 9.1 item would not be worth it as the trinket ilvls, and weapon ilvls far exceed jotuns value.