Parse coefficients from trigger data
A common motif found in recent spells, particularly damage procs from items & effects
is to have the damage coefficient determined by the drivers, rather than the damage
spell itself, with a dummy trigger on the damage spell referring back to the driver.

If ALL of the following are true:
  1) the special_effect_t is an offensive spell action or an attack action
  2) the proc action has 0 damage after parsing the spell data
  3) the proc action has a trigger effect referencing the driver
  4) the driver has a trigger effect referencing the proc action
then the coefficient of the effect in 4) will be used for the damage of action 2).