[Gear] Return functionality of cruciform_veinripper_proc_rate
Tanks have been using this option to mean "time spent behind boss," and not literally changing the Proc rate. This is how the variable used to work prior to the latest commit.

This returns the functionality of that variable to the state that the tank Discords are using it and makes it clearer. Perhaps the variable name should be changed as well?