[Druid] adjust adaptive swarm friendly swarm behavior
Friendly swarms will now potentially jump to 2 camps of virtual targets:
* `druid.adaptive_swarm_jump_distance_melee` sets the average distance of the melee camp
* `druid.adaptive_swarm_jump_distance_ranged` sets the average distance of the ranged camp
* `druid.adaptive_swarm_jump_distance_stddev` sets the standard deviation the virtual targets in each camps

When a damage swarm expires and jumps as a healing swarm, it was a 50% chance to jump to either camp.
When a healing swarm expires and jumps as a damage swarm, the travel distance will now match the randomized distance of the virtual friendly target.