[CASC] add script to help keep dbfile updated
This script parses the game exe to look for the structures used to reference GameTables and DB2s. When using this script, the resulting dbfile should be checked for false positives. False positives should never happen for DB2s due to the large number of constraints on the structure, but the apparent structure for GameTables is relatively simple and false positives can sometimes occur.