[Balance] update starfall
Starfall is not an asynchronous ticking buff.
* every second the buff tick calls the driver
* the driver has a random delay (up to 0.8s), randomized per target, before the projectile is launched
* the projectile's damage is snapshot at the time of launch based on the number of stacks of the buff
* there is tick_on_application, not tick_zero

This means that depending on the timing of the random delay and because it is tick_on_application,
it is possible for some targets to get the full 9 ticks of a stacked starfall and some targets will
lose a tick and only get 8 ticks worth of damage.