[DBC] handle entries on the same node with the same selection index
It's possible to have nodes with entries that have the same selection index.

In such cases, the game seems to assign the first choice to the entry that
comes earlier in TraitNodeXTraitNodeEntry.db2 itself.

As entries are added in this same order to _trait_nodes[node_id], we can
adjust here by incrementing the selection_index on successive entries with
the same selection_index as an already processed entry that is on the same
node.