[Havoc] Avoid wasting Inner Demon procs
The talent Inner Demon causes your first chaos strike or annihilation after eye beam or meta to do an additional hit of aoe damage. This proc was being overwritten during the opener and any other meta usages.

This change has us intentionally spend inner demon when meta or eyebeam are (going to be) ready. This results in a minor increase in both single target and aoe.