[Havoc] Avoid wasting Inner Demon procs (#7269)
* [Havoc] Avoid wasting Inner Demon procs

The talent Inner Demon causes your first chaos strike or annihilation after eye beam or meta to do an additional hit of aoe damage. This proc was being overwritten during the opener and any other meta usages.

This change has us intentionally spend inner demon when meta or eyebeam are (going to be) ready. This results in a minor increase in both single target and aoe.

* [Havoc] Use VR late in GCD

Currently, with spending ID procs in the opener, VR is being cast immediately after annihilation.

This change causes VR to be cast 0.4s before the next gcd at the earliest. By effectively lengthening the initiative window, at low haste levels outside of lust this results in a minor gain (~0.2%) and a negligible difference in more haste-rich scenarios.