[Warlock] Destruction APL Improvements (#7288)
* [Warlock] Havoc ST rule change

After last weeks hotfix, it's it not worth anymore to use havoc outside of infernal if burn to ashes is talented.

* [Warlock] fix trinket usage for 1:30min trinkets

Fix is mostly for shadowgem trinket. As it's a 1:30min trinket we can use it again between 2 infernals. 
Comparison : https://www.raidbots.com/simbot/report/d2PQBU6eygv2g5UvEhGjBa

* [Warlock] Soul Fire optimisation in ST

* [Warlock] SF cleave optimisation

* [Warlock] Prophylactic line for soulkeeper

It is currently dps neutral to cast soulkeeper in ST.
After changes to soulkeeper in 10.0.5, soulkeeper might be an increase to be casted.
This line is prophylactic for next patch.

* [Warlock] Destruction new profile

New talents for best profile

* [Warlock] race change for destruction