[Warlock] Update multitarget Immolate (#7359)
* New CB/RoF rules and Cry havoc in cleave

* RB rules and implement Cry Havoc

* trinkets

* [Warlock] Incinerate weaving

* [Warlock] Havoc ST rule change

After last weeks hotfix, it's it not worth anymore to use havoc outside of infernal if burn to ashes is talented.

* [Warlock] fix trinket usage for 1:30min trinkets

Fix is mostly for shadowgem trinket. As it's a 1:30min trinket we can use it again between 2 infernals. 
Comparison : https://www.raidbots.com/simbot/report/d2PQBU6eygv2g5UvEhGjBa

* [Warlock] Soul Fire optimisation in ST

* [Warlock] SF cleave optimisation

* [Warlock] Prophylactic line for soulkeeper

It is currently dps neutral to cast soulkeeper in ST.
After changes to soulkeeper in 10.0.5, soulkeeper might be an increase to be casted.
This line is prophylactic for next patch.

* [Warlock] Destruction new profile

New talents for best profile

* [Warlock] race change for destruction

* smarter rof

* [Warlock] Destro immolate + rof optimisation

* [Warlock] fix on target rof

* [Warlock] Destro immolate + rof optimisation

* adding soulkeeper in cleave apl

* fix spam incinerate in havoc

Fixing an issue where havoc apl spams incinerate if Rain of Fire is not talented and the number of enemies is bigger than the threshold (4-5)

* Fix pathing for aoe/cleave/st

Some cases, with more than 3 target, the apl was following ST logic. This should fix and force aoe apl in cases where there is 4 or more targets

* removing old havoc condition line

* [Warlock] Variable for tracking havoced immolate

* a bit more optimizations

* 2T bta stacks and madness