Implement player-scope expression support for temporary_enchant option
Temporary enchant option value now takes an additional if= option for a
given slot option. For example,
temporary_enchant=main_hand:howling_rune_3,if=!talent.improved_flametongue_weapon
will apply howling_rune_3 to the main hand, if the Improved Flametongue
Weapon talent is not selected.

If multiple temporary enchants for the same slot are given, the first
available one is used (i.e., either the if expression evaluates to
non-zero, or there is no if expression).

Note that due to the initialization ordering where temporary enchants
are applied to slots, not all player- and sim-scope options function
sensibly.