[Dot] adjust default arguments for dot_t::adjust_duration
Default behavior in-game when dots are adjust (duration is extended or reduced)
is to retain the snapshot state. Furthermore such adjustments are generally
not expected to count as a refresh for reporting purposes.

We change the default argument to reflect these assumptions:
1) snapshot state is not updated
2) the adjustment is not reported as a refresh

Snapshot updating can still be accomplished via passing `-1` as the state_flag.
NOTE: this will re-snapshot using the snapshot_flags and not the update_flags.

Following modules are adjusted to reflect the new default parameters:
        modified:   engine/class_modules/priest/sc_priest_discipline.cpp
        modified:   engine/class_modules/priest/sc_priest_shadow.cpp
        modified:   engine/class_modules/sc_druid.cpp
        modified:   engine/class_modules/sc_evoker.cpp
        modified:   engine/class_modules/sc_mage.cpp
        modified:   engine/class_modules/sc_rogue.cpp
        modified:   engine/class_modules/sc_shaman.cpp