temp fix Fote / CA clash
FotE delays KS empower substantionally, causing missed KS casts. in reality, often macros are used to micro the pet attacks.

it was always assumed that using the macros are a slight loss due to losing a certain quantity of bomb empowers, but this remains to be seen now. Will need to adapt the pet APL to give that a try at a later date.