[Enemy] Add enemy_custom_health_timeline option
This option can be used to adjust the time spent in various health
intervals. The user specifies health percentages at various points in
the iteration and the sim will interpolate the rest.

enemy_custom_health_timeline=pct1:time1/pct2:time2/...

pct is a value between 0 and 100, time is a value between 0 and 1, with
0 being start of the iteration and 1 being the end.