[Gear] Refactor & Fix Issue with Thorncaller Claw
Thorncaller claw's dummy debuff's refresh behaviour did not match the dot's refresh behaviour because it did not have a tick period/behaviour set. This caused thorn burst to not proc toward the end of thorn spirit if the dot was refreshed. Similarly, the dot would not spread to a new target on death if it had been refreshed.

My solution is to just remove the dummy debuff and use the dot's is_ticking() for the thorn burst proc and on_demise checks.