[Buff] add buff_t::consume( action_t* )
Buffs of the general scheme "your next X does/has Y" will match the action
against the buff's proc flags to check if the buff should be expired. This
often leads to the situation where the buff benefits the action but is not
expired because the flags do not match.

Calling buff_t::consume( action ) instead of buff_t::expire() will match
this behavior