[Buff] add buff_t::trigger( action_t* )
Passing an action_t* as the first argument to buff_t::trigger( ... )
will check the action against the proc flags of the buff_t::trigger_data
spell. Default buff_t constructor will set the trigger_data to be the
same as the buff spell, but this can (and should) be overriden via
buff_t::set_trigger_data().