[Player] add default gear loadout
`load_default_gear=1` player scoped option will load default gear on
player creation. Any items with explicit slot lines will override the
starting gear for that slot.

Default gear are starting boosted character items that are scaled to be
6 tiers higher than the heroic dungeon finder required item level.

For the initial implementation of The War Within, this value is:
    500 + 13 * 6 = 578