[Callback] implement dbc_proc_callback_t::activate_with_buff() helper function
Helper function to set up the callback to activate/deactivate based on
buff being triggered/expired.

NOTE: If the callback is created after player_t::init_special_effects(),
such as within a create_debuff() called on a new target_specific_debuff,
you MUST call this with the 2nd argument set to true to ensure the
callback is properly initialized.