[Gear] allow terser create_proc_action arguments
Previously create_proc_action only utilized the name argument as a
search parameter to see if the action already exists, and does not use
it for any further purpose. This means that for base classes that
require a name argument, the name must be provided again as a separate
argument to create_proc_action.

Now, if create_proc_action cannot construct the action with the given
arguments, it will attempt to construct by passing the search filter
name argument as the action class name argument, allowing e.g.:

    create_proc_action<generic_proc_t>( "name", effect, spell );

instead of:

    create_proc_action<generic_proc_t>( "name", effect, "name", spell );