Rename 'spell_speed' to 'spell_cast_speed' and 'attack_speed' to 'auto_attack_speed' (#8853)
'spell speed' is used for the combined effect of spell haste and the
much rarer spell cast speed effects. However, the similarly labled
'attack speed' and 'melee speed' apply only to auto attacks, which is
not exactly analogous to spell speed.

This further clarifies the distinction the two to make it clearer for
those unfamiliar with the specific SimC labels.

'spell cast speed' is used for the combined spell haste + spell cast
speed effects.

'auto attack speed' is used to refer only the auto attacks, which are
affected by the combined attack haste and auto attack speed effects.