[Parse] add CONSUME_BUFF parse flag
Passing CONSUME_BUFF to action-scoped parse_effects() with a buff will
set that buff to decrement() upon execute() of the action if that buff
has affected that action during any step up to the execute().

This only applies to effects that effect the action and are applied to
the player. Effects that apply to the target will not work.