[Consumables] implement The War Within foods
TWW foods can no longer be automatically parsed, as coefficents are
consolidated under spell 456961, and no longer present in the various
"well fed" buffs. The previously used indicator of 10s periodic trigger
now only points to 456961 and the only way to determine the specific
stat effect of the food is to manually examine the tooltip description
of the food item "refreshment" spell.

The HTML is also updated to display the item tooltip for all consumable
related entries where possible. This applies to all consumable actions
and to consumable buffs with custom implementations.