[RNG] cache gaussian cdf
Implement a gauss_t class for use with simple normal distributions with
only a mean and standard deviation, neither which are dynamic
(typically set via a sim option parameter). The minimum and maximum
cumulative distribution function are cached resulting resulting in
elimination of ~50% processing time compared to fully calculating the
gaussian every call.

A templated version of rng::gauss<MEAN, STDDEV>gauss() is also
implemented for compile-time constant mean and standard deviation.
The template parameters are given in milliseconds.