[Gear] elemental focusing lens initial implementation
NOTE: in-game seems to be using the coeff from the driver instead of the
equip, even though the tooltip refers to the coeff on the equip. The
driver's coeff is 10x the equip coeff, which is reflected in game. For
now it is implemented as if using the coeff from the equip as per tooltip.

// TODO: determine if it can proc with no gems
// TODO: determine which coeff is actually used
// TODO: confirm damage amount increase per gem type calculation
// TODO: confirm rppm decrease per gem type calculation
// TODO: determine if heal uses same coeff as damage
// TODO: confirm damage/heal depends on target type, and not triggering action type