[Druid] use derived class for bear berserk
Bear berserk, depending on talent, dynamically reduces the cooldown of frenzied
regeneration to 0, and also resets the cooldown of growl, mangle, and thrash.

This reset needs to happen before the buff is applied to ensure that dynamic
cooldown adjustments is not called, as there is nothing to adjust when cooldowns
are forcibly reset as such.