[Parse] pass action & state to modified_spelleffect_t conditional functor
Conditions for whether to apply a particular modifier to a modified_spelleffect_t
can be based on the current action or action_state that is being processed.
The various spelleffect_t analogues such as base_value() and resource() can
be called with action & action_state pointers as arguments to pass these
along to the functor.