[shaman] implement Ancient Fellowship as deck-of-cards
In a recent blue post[^1] Blizzard suggested that Ancient Fellowship
would become less random. Given the datamined changes[^2] to the spell
including the addition of two effects, which combined can form the
expected percentage to proc the effect, we believe this was done in a
deck-of-cards fashion similar to Icefury's rework.

This change pretty much just copies the Icefury deck implementation
directly, including support for class options to override the default
values.

[^1]: https://us.forums.blizzard.com/en/wow/t/the-war-within-beta-development-notes/1870426/15
[^2]: https://www.wowhead.com/news/datamined-class-and-spell-changes-on-war-within-beta-build-55789-lots-of-class-345433