[shaman] give fusion of elements buffs semantic names (#9099)
I think it's expected that at some point in the future we may want
conditional behavior in the APL based upon the state of Fusion of
Elements buffs. While the current names for these buffs will function,
remembering which buff is for nature and which is for fire by numeric
mappings is bound to cause a mistake at some point.

To that end, this change simply renames them to
`fusion_of_elements_nature` and `fusion_of_elements_fire` so it is
immediately obvious which is which.