[shaman] Update EQ SP Coefficient for TWW buffs
Recent hotfixes reset the scaling effects on the elemental
specialization aura to 0% and moved buffs baseline into abilities.

As a result Earthquake's spell power modifier is actually being
increased, while still reflecting the overall nerf of 15% in-game.

While #9137 is in review, I wanted to get this change out so at least
Earthquake numbers will be accurate (on Beta).