Remove get_background_action with non-unique action names
In the hunter module, get_background_action is often used with the presumable expectation to only create a single action per hunter (purely as an optimization). Since the background actions do not have pass on their constructor name to the action::name, that is not the case and the action get's constructed every time anyway. At the same time, the constructor name are just silently hidden instead to conform to the get_background_action.

Add an assert to the function, remove these non-intuitive usages.