[shaman] Fix Ancestor Elemental Blast casts
It seems a recent spell data change resulted in simc considering this
spell power coefficient to be 0. This seems to roughly align with the
Ancestor Elemental Blast spell (447427) receiving a change that resulted
in its tooltop showing a `0` value rather than the spell power
coefficient.

Thankfully it still has a Dummy effect holding the correct spell power
coefficient, so we can use that with an override.