[shaman] Only harmful spells proc Fusion of Elements
During some APL examination, we discovered that Ancestral Swiftness was
consuming buffs for Fusion of Elements while it does not do so in-game.

There is technically one incorrect behavior induced by this change,
which is that Wind Shear (interrupt) _does_ trigger Fusion of Elements
on Beta, but being as it's the only spell to do so, we believe this to
be a bug that will get fixed.