[shaman] Allow background casts to consume Master of the Elements (#9165)
Somewhat counterintuitively, background casts of spells like Flame Shock do
consume this buff in-game (eg assume MOTE is proc'd, casting Primordial
Wave does not consume MOTE, and the Flame Shock it triggers does).

It's likely we set this to never proc on-background on the assumption we
wanted to avoid overloads eating this proc. However, since the proc is
consumed on execute, and the base spell execute for abilities like
Lightning Bolt and Chain Lightning will still consume this before their
respective overloads.

It's unclear if this will brick some other assumptions elsewhere in the
module so I'm leaving this as a draft for now.