[Shaman] adjust for using proper find_talent_spell to find ascendance spell data
Previously the old grid based talent system was used to find spell data for
ascendance. A proxy `spell.ascendance` is now used and will contains the
spell from the talent lookup if Ascenance is talented, or hard coded spell
id lookup if Ascendance is not talented but Deeply Rooted Elements is.