[shaman] Hotfix Stormkeeper to be 3 max stacks
In an undocumented change with the most recent round of hotfixes,
Stormkeeper was hotfixed to allow for a maximum of 3 stacks, and
allowing Rolling Thunder to activate while Stormkeeper is already up (or
vice versa) without wasting any stacks of the buff.

Looking at this more closely, @HawkCorrigan found that these stack
changes don't seem to be in the spell database, or at the very least,
not automatagically plumbed in.

To work around this, use a hotfix to override the max stacks, and then
specify an initial stack count to avoid current calls to `trigger()`
yielding more stacks than intended.