[Shaman] Introduce experimental Spirit Wolf-related expressions
`wolves_active_for.<sub-expression>` expression evaluates to the number
of active Spirit Wolves the actor has within the time-period defined by
the sub-expression. The sub-expression may be any valid Simulationcraft
expression. For example, the expression
`wolves_active_for.buff.ascendance.remains` would return the number of
active Spirit Wolves within the remaining duration of the Ascendaance
buff. The expression is defined in the action scope, so an
expression such as `wolves_active_for.gcd` would evaluate to the number
of wolves active during the Global Cooldown duration of the APL line the
expression is used in.

Similar to `wolves_active_for`, the `wolves_expiring_in.<subexpression>`
will evaluate to the number of Spirit Wolves that have less remaining
duration than the time evaluated by the subexpression.

Note that any and all expressions are valid, as Simulationcraft does not
separate between time-based or "value based" expressions.