[Buff] implement buff_t::consume() and proper icd handling
An action consuming a buff (as opposed to triggering a buff) checks against
the buff data's proc flags and the buff data's internal cooldown (which blizzard
sometimes calls the "spell proc cooldown"). The analogues for when the buff
is triggered is usually found on the driver's spell data.

Currently SimC has the following incorrect behavior:
* Consuming a buff is done via buff_t::decrement() or buff_t::expire(), which
does not account for the buff data's ICD and unless using the action_t* argument
override, does not account for the buff data's proc flags.
* The buff data's ICD can be parsed as the buff's cooldown, and thus apply
against the triggering of the buff instead of consuming the buff.
* buff_t::set_trigger_spell() does not parse the ICD of the new trigger spell
data to apply to the buff's ICD.

The following is implement to correct this:
* buff_t::consume( action_t*. [int] ) can be used to check consumption of the buff.
It will check the action against the buff's proc flags as well as the ICD.
By default, ALL STACKS will be consumed and the buff expire()'d. An optional
[int] value can be passed to specify how many stacks to consume.
* buff_t::set_consume_all_stacks( bool ) can be used to change the default
consume() behavior. When set false, only ONE STACK will be consumed.
* buff_t::set_trigger_spell() will now properly parse the ICD of the trigger
spell data and apply it to the buff's cooldown. This means that the ICD of
the triggering spell will become the ICD applied when the buff is triggered.
* buff_t::apply_affecting_aura() will now properly parse flat modifiers to
P_PROC_COOLDOWN ("spell proc cooldown") and apply it to the buff's ICD.
* parse_effects() flags of EXPIRE_BUFF & DECREMENT_BUFF have been consolidated
into CONSUME_BUFF.
* buff_t::can_expire() has been renamed buff_t::can_consume().
* "buff.internal_cooldown_remains" expression added.