[Action] pool_resource adjustments
* Default wait time is set to `player_t::available()` since `pool_resource`
indicates continuous re-evaluate of action priorities without going further
down the list past the pool_resource line.

* `pool_resource`` will now correctly account for the `extra_amount`` option by
pooling even if the next action is ready if current resource isn't sufficient
to cover cost + extra_amount.

* `force_wait` option can be set for force suspension of APL evaluation until
sufficent resources are available to cast the next action. This is in contrast
to standard behavior where APL evaluation is restarted when you hit the pool_resource
line. This is equivalent to a `wait` action that dynamically uses the cost
of the next action to determine how long to wait. This option requires `for_next`.

* Added some comments around next_action evaluation.